move jdbc dao to separate package
