+++ library/src/main/java/cyclone/otusspring/library/dao/jdbc/BookDaoJdbc.java
@@ -1,5 +1,6 @@
-package cyclone.otusspring.library.dao;
+package cyclone.otusspring.library.dao.jdbc;
 
+import cyclone.otusspring.library.dao.BookDao;
 import cyclone.otusspring.library.dto.BookDetails;
 import cyclone.otusspring.library.model.Author;
 import cyclone.otusspring.library.model.Book;
