+++ library/src/main/java/cyclone/otusspring/library/dao/jdbc/GenreDaoJdbc.java
@@ -1,5 +1,6 @@
-package cyclone.otusspring.library.dao;
+package cyclone.otusspring.library.dao.jdbc;
 
+import cyclone.otusspring.library.dao.GenreDao;
 import cyclone.otusspring.library.model.Genre;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
